fix: reinstate nhp_aci imports to top
With the valuable help of YiWen, we noticed that one more item that could be automated is updating nhp-aci = { git = 'https://github.com/the-strategy-unit/nhp_aci.git' } automatically. This can become part of our pipeline, which is the next GitHub issue we'll tackle. Updating nhp_aci with 'uv install -e '.[dev]' '. Also, nhp-aci only looks for env vars in the current repo. For increased security, my .env lives outside the repo, which breaks nhp-aci. I'll open a PR to suggest a fix